split frames into files
