+++ Datas/Perso/1.json
@@ -16,7 +16,7 @@
   "constitution": 2,
   "sagesse": 2,
   "vitesse": 2,
-  "pdv": 67,
+  "pdv": 52,
   "armes": [
     {
       "name": "Epee courte",
@@ -31,17 +31,16 @@
   ],
   "inventaire": [
     {
-        "name": "potion",
-        "amount": 1
+      "name": "potion",
+      "amount": 1
     },
     {
-        "name": "super-potion",
-        "amount": 1
+      "name": "super-potion",
+      "amount": 1
     },
     {
-        "name": "mega-potion",
-        "amount": 1
+      "name": "mega-potion",
+      "amount": 1
     }
-  ],
-  "budget" : 30
+  ]
 }
