+++ Datas/Perso/test.json
@@ -16,7 +16,7 @@
   "constitution": 2,
   "sagesse": 2,
   "vitesse": 2,
-  "pdv": 47,
+  "pdv": 6,
   "armes": [
     {
       "name": "Epee courte",
