+++ src/Window/fight.py
@@ -1,0 +1,15 @@
+import json
+import time
+from tkinter import *
+import pygame
+import os
+import tkinter as tk
+from random import *
+from functools import partial
+
+from src.Combat.combat import Combat
+from src.Inventory.inventory import Inventory
+from src.Perso.person import Person
+from src.Window.quest import start_quest_frame
+
+
